# mímir/api/openapi.yaml
openapi: 3.2.0
info:
  title: Mímir Financial Data API
  version: 1.0.0
  summary: Gateway to financial analytics and modeling results
  description: |
    Mímir is a gateway to the Logos PostgreSQL database. It exposes
    financial analytics (IFRS statements and ratios) and financial
    modeling results over an HTTP REST API.

    Authentication is performed with mutual TLS (mTLS): clients present an
    X.509 certificate signed by the service's trusted CA.
servers:
  - url: https://localhost:8443
security:
  - clientCertificate: []
tags:
  - name: financials
    description: IFRS financial statements and ratios
  - name: industries
    description: Industry classification
  - name: companies
    description: Companies
  - name: scenarios
    description: What-if scenarios for forecasting
  - name: forecasts
    description: Forecasted financial results
paths:
  /financials:
    post:
      tags:
        - financials
      operationId: uploadFinancialData
      summary: Upload IFRS financial data for a company
      description: |
        Creates or updates a company and stores its absolute financial
        metrics and computed ratios for one or more fiscal years.

        The request body mirrors the arguments of the Logos
        `add_financial_data` function: a company, its industry, a mapping of
        metric code -> (year -> value), and a mapping of
        year -> (ratio name -> value).

        Values set to `null` are skipped. The industry must already exist in
        the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialDataUpload"
            examples:
              sberbank:
                summary: Single company, two fiscal years
                value:
                  companyName: "ПАО Сбербанк"
                  ticker: "SBER"
                  industryName: "Banks"
                  metrics:
                    Revenue:
                      "2023": 3500000000000
                      "2024": 3900000000000
                    "Net profit":
                      "2023": 1490000000000
                      "2024": null
                  ratios:
                    "2023":
                      "Current ratio": 1.5
                      "ICR": 3.2
                    "2024":
                      "Current ratio": 1.6
                      "ICR": 3.4
      responses:
        "201":
          description: Financial data accepted and stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialDataUploadResult"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Referenced industry does not exist in the database.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /industries:
    get:
      tags:
        - industries
      operationId: listIndustries
      summary: List industries
      description: Returns all industries.
      responses:
        "200":
          description: List of industries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Industry"
    post:
      tags:
        - industries
      operationId: createIndustry
      summary: Create an industry
      description: Creates a new industry classification node.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndustryCreate"
      responses:
        "201":
          description: Industry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Industry"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Industry code already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /industries/{industryId}:
    patch:
      tags:
        - industries
      operationId: updateIndustry
      summary: Update an industry
      description: Updates fields of an existing industry.
      parameters:
        - name: industryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndustryUpdate"
      responses:
        "200":
          description: Updated industry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Industry"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Industry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Industry code already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - industries
      operationId: deleteIndustry
      summary: Delete an industry
      description: Deletes an industry. Fails if it still has children or companies.
      parameters:
        - name: industryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Industry deleted.
        "404":
          description: Industry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Industry still has child industries or companies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies:
    get:
      tags:
        - companies
      operationId: listCompanies
      summary: List companies
      description: Returns companies, optionally filtered by industry or name.
      parameters:
        - name: industryId
          in: query
          required: false
          schema:
            type: integer
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of companies.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      tags:
        - companies
      operationId: createCompany
      summary: Create a company
      description: Creates a new company linked to an industry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyCreate"
      responses:
        "201":
          description: Company created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Referenced industry does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies/{companyId}:
    patch:
      tags:
        - companies
      operationId: updateCompany
      summary: Update a company
      description: Updates fields of an existing company.
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Referenced industry does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - companies
      operationId: deleteCompany
      summary: Delete a company
      description: Deletes a company and its financial data.
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Company deleted.
        "404":
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /scenarios:
    get:
      tags:
        - scenarios
      operationId: listScenarios
      summary: List scenarios
      description: Returns what-if scenarios, optionally filtered by active status.
      parameters:
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of scenarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
  /scenarios/{scenarioId}/variables:
    get:
      tags:
        - scenarios
      operationId: listScenarioVariables
      summary: List variables of a scenario
      description: Returns the metric adjustments that define a scenario.
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scenario variables.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScenarioVariable"
        "404":
          description: Scenario not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies/{companyId}/financials:
    get:
      tags:
        - financials
      operationId: getCompanyFinancials
      summary: Get historical financial data for a company
      description: Returns the company's stored financial metrics and ratios.
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Historical financial data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyFinancialData"
        "404":
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies/{companyId}/forecasts:
    put:
      tags:
        - forecasts
      operationId: uploadForecasts
      summary: Save forecast results for a company
      description: |
        Upserts forecasted metric values for a company and forecast year.
        Each entry references a scenario by code and metrics by code.
        Unknown metric codes are skipped.
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForecastUpload"
      responses:
        "201":
          description: Forecasts stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastUploadResult"
        "400":
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Company or scenario not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    clientCertificate:
      type: mutualTLS
      description: |
        Mutual TLS authentication. The client presents an X.509 certificate
        that must be signed by the CA configured on the server.
  schemas:
    FinancialDataUpload:
      type: object
      description: |
        IFRS financial data for a single company. Mirrors the arguments of
        the Logos `add_financial_data` function.
      required:
        - companyName
        - ticker
        - industryName
        - metrics
        - ratios
      properties:
        companyName:
          type: string
          description: Full name of the company.
          example: "ПАО Сбербанк"
        ticker:
          type: string
          description: Exchange ticker symbol.
          example: "SBER"
        industryName:
          type: string
          description: Name of the industry. Must already exist in the database.
          example: "Banks"
        metrics:
          $ref: "#/components/schemas/MetricSeries"
        ratios:
          $ref: "#/components/schemas/RatioSeries"

    MetricSeries:
      type: object
      description: |
        Absolute financial metrics (P&L, balance sheet, cash flow) keyed by
        metric code. Each value maps a fiscal year to the metric value.
        Null values are skipped.
      additionalProperties:
        type: object
        description: Metric values keyed by fiscal year.
        additionalProperties:
          type: [number, "null"]
          description: Metric value for the fiscal year.
      example:
        Revenue:
          "2023": 3500000000000
          "2024": 3900000000000
        "Net profit":
          "2023": 1490000000000

    RatioSeries:
      type: object
      description: |
        Computed ratios keyed by fiscal year. Each value maps a ratio name
        to its value. Null values are skipped.
      additionalProperties:
        type: object
        description: Ratio values keyed by ratio name.
        additionalProperties:
          type: [number, "null"]
          description: Ratio value for the fiscal year.
      example:
        "2023":
          "Current ratio": 1.5
          "ICR": 3.2
        "2024":
          "Current ratio": 1.6

    FinancialDataUploadResult:
      type: object
      description: Summary of a successful financial data upload.
      required:
        - companyId
        - companyName
        - ticker
        - metricCount
        - ratioCount
      properties:
        companyId:
          type: integer
          description: Database identifier of the company.
        companyName:
          type: string
          description: Full name of the company.
        ticker:
          type: string
          description: Exchange ticker symbol.
        metricCount:
          type: integer
          description: Number of metric values stored.
        ratioCount:
          type: integer
          description: Number of ratio values stored.

    Industry:
      type: object
      description: An industry classification node.
      required:
        - id
        - name
        - code
      properties:
        id:
          type: integer
          description: Database identifier.
        name:
          type: string
          description: Industry name.
        code:
          type: string
          description: Unique industry code.
        parentId:
          type: [integer, "null"]
          description: Parent industry id, or null for a top-level industry.

    IndustryCreate:
      type: object
      description: Payload for creating an industry.
      required:
        - name
        - code
      properties:
        name:
          type: string
          description: Industry name.
        code:
          type: string
          description: Unique industry code.
        parentId:
          type: [integer, "null"]
          description: Parent industry id, or null for a top-level industry.

    IndustryUpdate:
      type: object
      description: Payload for partially updating an industry.
      properties:
        name:
          type: string
          description: Industry name.
        code:
          type: string
          description: Unique industry code.
        parentId:
          type: [integer, "null"]
          description: Parent industry id, or null for a top-level industry.

    Company:
      type: object
      description: A company linked to an industry.
      required:
        - id
        - name
        - industryId
      properties:
        id:
          type: integer
          description: Database identifier.
        ticker:
          type: [string, "null"]
          description: Exchange ticker symbol.
        name:
          type: string
          description: Company name.
        inn:
          type: [string, "null"]
          description: Russian tax identification number (INN).
        industryId:
          type: integer
          description: Id of the industry the company belongs to.

    CompanyCreate:
      type: object
      description: Payload for creating a company.
      required:
        - name
        - industryId
      properties:
        ticker:
          type: [string, "null"]
          description: Exchange ticker symbol.
        name:
          type: string
          description: Company name.
        inn:
          type: [string, "null"]
          description: Russian tax identification number (INN).
        industryId:
          type: integer
          description: Id of the industry the company belongs to.

    CompanyUpdate:
      type: object
      description: Payload for partially updating a company.
      properties:
        ticker:
          type: [string, "null"]
          description: Exchange ticker symbol.
        name:
          type: string
          description: Company name.
        inn:
          type: [string, "null"]
          description: Russian tax identification number (INN).
        industryId:
          type: integer
          description: Id of the industry the company belongs to.

    CompanyFinancialData:
      type: object
      description: Historical financial metrics and ratios of a company.
      required:
        - companyId
        - companyName
        - metrics
        - ratios
      properties:
        companyId:
          type: integer
          description: Database identifier of the company.
        companyName:
          type: string
          description: Full name of the company.
        ticker:
          type: [string, "null"]
          description: Exchange ticker symbol.
        metrics:
          $ref: "#/components/schemas/MetricSeries"
        ratios:
          $ref: "#/components/schemas/RatioSeries"

    Scenario:
      type: object
      description: A what-if scenario used for forecasting.
      required:
        - id
        - code
        - name
        - category
        - isActive
      properties:
        id:
          type: integer
          description: Database identifier.
        code:
          type: string
          description: Unique scenario code.
        name:
          type: string
          description: Scenario name.
        category:
          type: [string, "null"]
          description: Scenario category.
        description:
          type: [string, "null"]
          description: Human-readable scenario description.
        isActive:
          type: boolean
          description: Whether the scenario is active.

    ScenarioVariable:
      type: object
      description: A metric adjustment within a scenario.
      required:
        - metricCode
        - operator
        - value
      properties:
        metricCode:
          type: string
          description: Code of the metric the adjustment applies to.
        operator:
          type: string
          description: How the adjustment is applied to the metric.
          enum:
            - multiply
            - add
            - set
            - scale_to_revenue
        value:
          type: number
          description: Adjustment value.

    ForecastEntry:
      type: object
      description: Forecasted metric values for one scenario.
      required:
        - scenarioCode
        - metrics
      properties:
        scenarioCode:
          type: string
          description: Scenario code the forecast belongs to.
        metrics:
          type: object
          description: Metric code -> forecast value.
          additionalProperties:
            type: number

    ForecastUpload:
      type: object
      description: Forecast results to store for a company.
      required:
        - forecastYear
        - scenarios
      properties:
        forecastYear:
          type: integer
          description: Forecasted fiscal year.
        scenarios:
          type: array
          description: Forecasted values per scenario.
          items:
            $ref: "#/components/schemas/ForecastEntry"

    ForecastUploadResult:
      type: object
      description: Summary of a successful forecast upload.
      required:
        - companyId
        - forecastYear
        - scenarioCount
        - metricCount
      properties:
        companyId:
          type: integer
          description: Database identifier of the company.
        forecastYear:
          type: integer
          description: Forecasted fiscal year.
        scenarioCount:
          type: integer
          description: Number of scenarios stored.
        metricCount:
          type: integer
          description: Number of forecast values stored.

    Error:
      type: object
      description: Standard error response.
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error message.